Move bar chart out of activity.
